🔨 [Refactor] ProfileInfo 클래스를 별도의 파일로 분리하여 관리 용이성 향상
